# src/documentation/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Swagger Javascript Fullstack API",
    "description": "Documentación API: Juego Javascript Fullstack"
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users": {
      "get": {
        "description": "Devuelve todos los usuarios.\n",
        "parameters": [],
        "operationId": "findUsers",
        "responses": {
          "200": {
            "description": "Devuelve todos los usuarios de la aplicación.\n",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            },
            "500": {
              "description": "Devuelve un error",
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      },
      "post": {
        "description": "Añade un nuevo usuario al juego",
        "operationId": "addUser",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "Añade un nuevo usuario al juego",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve el nuevo usuario creado.",
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          },
          "500": {
            "description": "Muestra un error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Devuelve un usuario en función de su ID",
        "operationId": "find user by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del usuario buscado",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve en usuario encontrado mediante ID.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "No se ha encontrado ningún usuario",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Se ha producido un error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "patch": {
        "description": "Actualiza un usuario.",
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del usuario a actualizar",
            "required": true,
            "type": "string"
          },
          {
            "name": "user",
            "in": "body",
            "description": "Usuario actualizado",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve el usuario actualizado.",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "description": "Elimina un usuario en función de su ID",
        "operationId": "deleteUser",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID del usuario a eliminar",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Return el usuario borrado",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "404": {
            "description": "No encontrado",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/games": {
      "get": {
        "description": "Devuelve todos los juegos.\n",
        "parameters": [],
        "operationId": "findGames",
        "responses": {
          "200": {
            "description": "Devuelve todos los juegos.\n",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Game"
              }
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "description": "Crea una nueva partida",
        "operationId": "addGame",
        "parameters": [
          {
            "name": "game",
            "in": "body",
            "description": "Partida para añadir al juego",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewGame"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return partida.",
            "schema": {
              "$ref": "#/definitions/NewGame"
            }
          },
          "500": {
            "description": "error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/games/{id}": {
      "get": {
        "description": "Devuelve un juego mediante su ID",
        "operationId": "find game by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of game to fetch",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve el juego.",
            "schema": {
              "$ref": "#/definitions/Game"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "Error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "patch": {
        "description": "Actualiza la partida.",
        "operationId": "updateGame",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de la partida a actualizar",
            "required": true,
            "type": "string"
          },
          {
            "name": "game",
            "in": "body",
            "description": "Game to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewGame"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Devuelve el juego actualizado.",
            "schema": {
              "$ref": "#/definitions/Game"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "General error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "description": "deletes a game based on the ID supplied",
        "operationId": "deleteGame",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of game to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns game deleted",
            "schema": {
              "$ref": "#/definitions/Game"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "General error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/rooms": {
      "get": {
        "description": "Returns all rooms.\n",
        "parameters": [],
        "operationId": "findRooms",
        "responses": {
          "200": {
            "description": "Returns all rooms from the application",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Room"
              }
            }
          },
          "500": {
            "description": "General error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "post": {
        "description": "Creates a new room in the application.",
        "operationId": "addRoom",
        "parameters": [
          {
            "name": "room",
            "in": "body",
            "description": "Room to add to the application",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewRoom"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns new room.",
            "schema": {
              "$ref": "#/definitions/NewRoom"
            }
          },
          "500": {
            "description": "General error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/rooms/{id}": {
      "get": {
        "description": "Returns a room based on a single ID",
        "operationId": "find room by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of room to fetch",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns room found.",
            "schema": {
              "$ref": "#/definitions/Room"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "General error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "patch": {
        "description": "Update a room in the application.",
        "operationId": "updateRoom",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of room to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "room",
            "in": "body",
            "description": "Room to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewRoom"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns updated room.",
            "schema": {
              "$ref": "#/definitions/Room"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "General error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "delete": {
        "description": "deletes a room based on the ID supplied",
        "operationId": "deleteRoom",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of room to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Returns room deleted",
            "schema": {
              "$ref": "#/definitions/Room"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          },
          "500": {
            "description": "General error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "Mongo Object Id"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "description": "Email format"
        },
        "password": {
          "type": "string"
        },
        "color":  {
          "type": "string"
        },
        "image": {
          "type": "string",
          "description": "Image in base64"
        }
      }
    },
    "NewUser": {
      "type": "object",
      "required": [
        "name",
        "email",
        "password",
        "color"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "description": "Valid email format"
        },
        "password": {
          "type": "string"
        },
        "color":  {
          "type": "string",
          "description": "Hexadecimal color"
        },
        "image": {
          "type": "string",
          "description": "Base64 image"
        }
      }
    },
    "PlayerData": {
      "type": "object",
      "properties": {
        "playerId": {
          "type": "string"
        },
        "points": {
          "type": "number"
        },
        "state": {
          "type": "string",
          "enum": ["La partida aún no ha comenzado", "En progreso", "Salir"]
        }
      }
    },
    "Game": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "Mongo ObjectId"
        },
        "playersData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PlayerData"
          }

        },
        "turn": {
          "type": "string",
          "description": "Player ID (Mongo ObjectId) who owns the turn"
        }
      }
    },
    "NewGame": {
      "type": "object",
      "required": [
        "playersData"
      ],
      "properties": {
        "playersData": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/PlayerData"
          }

        },
        "turn": {
          "type": "string",
          "description": "Player ID (Mongo ObjectId) who owns the turn"
        }
      }
    },
    "Room": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string",
          "description": "Mongo ObjectId"
        },
        "name": {
          "type": "string"
        },
        "state": {
          "type": "string",
          "enum": ["La partida aún no ha comenzado", "En curso"]
        },
        "game": {
          "type": "string",
          "description": "Mongo ObjectId"
        },
        "users": {
          "type": "array",
          "items": {
            "type": "string",
            "description": "User Mongo ObjectId"
          }
        }
      }
    },
    "NewRoom": {
      "type": "object",
      "required": [
        "name",
        "state"
      ],
      "properties": {
        "name": {
          "type": "string"
        },
        "state": {
          "type": "string",
          "enum": ["La partida aún no ha comenzado", "En curso"]
        },
        "game": {
          "type": "string",
          "description": "Mongo ObjectId"
        },
        "users": {
          "type": "array",
          "items": {
            "type": "string",
            "description": "User Mongo ObjectId"
          }
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "status": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  }
}